requirements.txt
Django==4.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
drf-yasg==1.21.4
drf-spectacular==0.26.3
django-environ==0.10.0
django-filter==23.2
python-decouple
